Migration Complete
All runner code migrated to shared class, GUI is still working.  Need to modify DataBuilder to NOT read from User Settings when using Console.
