+++ Snap2HTML-NG.Shared/Snap2HTML-NG.Shared.csproj
@@ -33,6 +33,8 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Drawing" />
+    <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -41,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Builder\DataBuilder.cs" />
     <Compile Include="Models\CommandLineModel.cs" />
     <Compile Include="Models\SnappedFile.cs" />
     <Compile Include="Models\SnappedFolder.cs" />
@@ -50,9 +53,7 @@
     <Compile Include="Utils\Legacy\CommandLine\Arguments.cs" />
     <Compile Include="Utils\Legacy\Helpers.cs" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Helpers\" />
-  </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
     <Content Include="template.html">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
